split hybrid_window.cpp
