+++ inst/include/dplyr/HybridHandlerMap.h
@@ -10,6 +10,7 @@ void install_simple_handlers(HybridHandlerMap& handlers);
 void install_minmax_handlers(HybridHandlerMap& handlers);
 void install_count_handlers(HybridHandlerMap& handlers);
 void install_nth_handlers(HybridHandlerMap& handlers);
+void install_window_handlers(HybridHandlerMap& handlers);
 
 bool hybridable(RObject arg);
